Move sun condition tests
